added product details page to application.
